feat: make more pretty add links message, fix bugs

+++ functions/linkMessage/callbacks.js
@@ -15,12 +15,19 @@ const callbackDeleteLink = async (callbackQuery) => {
     const messageId = callbackQuery.message.message_id;
     const choosenOption = callbackQuery.data;
 
+    if (
+        choosenOption === 'deleteLink_deleteYes' ||
+        choosenOption === 'deleteLink_deleteNo' ||
+        /^addLinks_link/.test(choosenOption) ||
+        choosenOption === 'addLinks_cancel' ||
+        choosenOption === 'addLinks_deleteAllLinks' ||
+        choosenOption === 'addLinks_accept'
+    ) return;
+
+
     await bot.deleteMessage(chatId, messageId);
 
-    if (choosenOption === 'allLinkDeleteYes' || choosenOption === 'allLinkDeleteNo' || choosenOption === 'cancel') {
-        return;
-    }
-
+    if (choosenOption === 'deleteLink_cancel') return;
     await deleteHw(choosenOption);
     await renderLinkMessage(chatId);
 }
@@ -29,7 +36,8 @@ const callbackDeleteAllLink = async (callbackQuery) => {
     const userChoise = callbackQuery.data;
     const chatId = callbackQuery.message.chat.id;
 
-    if (userChoise === "allLinkDeleteYes") {
+    if (userChoise === "deleteLink_deleteYes") {
+        await bot.deleteMessage(chatId, messageId);
         await deleteAllLinksFromServer();
         await renderLinkMessage(chatId);
     }
